requirements.txt
asgiref==3.9.2
black==25.9.0
build==1.3.0
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.12.0
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pip-tools==7.5.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.2.0
ruff==0.14.4
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
